requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
cryptography==44.0.1
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
Flask-Session==0.8.0
Flask-WTF==1.2.2
flask-limiter
redis
Flask-SocketIO
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
msgspec==0.19.0
ninja==1.11.1.3
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
threadpoolctl==3.5.0
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
torch
torchaudio
